Signature should be based on lowercase url with sorted query params.
Query params are sorted by name and the entire url is lowercased before
generating the signature. Defining a consistent methodology for
converting a URL with query params into a signature minimizes the
possibility that there will be a problem verifying the signature. The
method outlined above is the one recommended by OAuth for this reason.